requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyJWT[crypto]
cryptography
